VoicesPopup to not cover cancel button
Configure the popup for voices in the Parts menu to open upwards if
opening downwards means that it will cover the cancel/open buttons.
Achieved this by using a function in PartsView to determine whether
the popup should open upwards. Another function was added in
PartDelegate so that, should the popup open upwards, its positioning
would be correct.